requirements.txt
Django==2.0.7
django-polymorphic
